functioning track end points

+++ audicin-track-api/src/main/java/com/audicin/backend/track/api/db/repositories/LicenseRepository.java
@@ -4,6 +4,6 @@ import com.audicin.backend.track.api.db.models.License;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LicenseRepository extends JpaRepository<License, Long> {
+public interface LicenseRepository extends JpaRepository<License, Integer> {
     List<License> findByTrackId(Long trackId);
 }
